Renamed packages to avoid confusion wiht old oslc-java-client
