+++ oslc4j-client/pom.xml
@@ -80,7 +80,7 @@
         <dependency>
             <groupId>javax.ws.rs</groupId>
             <artifactId>javax.ws.rs-api</artifactId>
-            <version>2.1.1</version>
+            <version>2.1</version>
         </dependency>
         <!--TEST-->
         <dependency>
@@ -155,11 +155,6 @@
                 <groupId>org.apache.httpcomponents</groupId>
                 <artifactId>httpclient</artifactId>
                 <version>4.5.6</version>
-            </dependency>
-            <dependency>
-                <groupId>org.apache.httpcomponents</groupId>
-                <artifactId>httpcore</artifactId>
-                <version>4.4.10</version>
             </dependency>
             <dependency>
                 <groupId>com.ibm.icu</groupId>
