+++ oslc4j-client/src/main/java/org/eclipse/lyo/client/OSLCConstants.java
@@ -16,7 +16,7 @@
  *    Samuel Padgett  - deprecate constants for non-existent terms
  *******************************************************************************/
 
-package org.eclipse.lyo.client.oslc;
+package org.eclipse.lyo.client;
 
 /**
  * General OSLC constants
