+++ oslc4j-client/src/main/java/org/eclipse/lyo/client/resources/AutomationPlan.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2012 IBM Corporation.
+ * Copyright (c) 2012, 2014 IBM Corporation.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -11,9 +11,10 @@
  *
  * Contributors:
  *
- *     Paul McMahan         - initial API and implementation
+ *     Paul McMahan <pmcmahan@us.ibm.com>        - initial implementation
+ *     Samuel Padgett <spadgett@us.ibm.com>      - fix getParameterDefinitions annotations
  *******************************************************************************/
-package org.eclipse.lyo.client.oslc.resources;
+package org.eclipse.lyo.client.resources;
 
 import java.net.URI;
 import java.util.Arrays;
@@ -23,10 +24,12 @@ import java.util.TreeSet;
 
 import org.eclipse.lyo.oslc4j.core.annotation.OslcDescription;
 import org.eclipse.lyo.oslc4j.core.annotation.OslcName;
+import org.eclipse.lyo.oslc4j.core.annotation.OslcNamespace;
 import org.eclipse.lyo.oslc4j.core.annotation.OslcOccurs;
 import org.eclipse.lyo.oslc4j.core.annotation.OslcPropertyDefinition;
 import org.eclipse.lyo.oslc4j.core.annotation.OslcRange;
 import org.eclipse.lyo.oslc4j.core.annotation.OslcReadOnly;
+import org.eclipse.lyo.oslc4j.core.annotation.OslcResourceShape;
 import org.eclipse.lyo.oslc4j.core.annotation.OslcTitle;
 import org.eclipse.lyo.oslc4j.core.annotation.OslcValueType;
 import org.eclipse.lyo.oslc4j.core.model.AbstractResource;
@@ -34,40 +37,79 @@ import org.eclipse.lyo.oslc4j.core.model.Occurs;
 import org.eclipse.lyo.oslc4j.core.model.OslcConstants;
 import org.eclipse.lyo.oslc4j.core.model.ValueType;
 
+@OslcResourceShape(title = "Automation Plan Resource Shape", describes = AutomationConstants.TYPE_AUTOMATION_PLAN)
+@OslcNamespace(AutomationConstants.AUTOMATION_NAMESPACE)
 /**
- * @see http://open-services.net/bin/view/Main/QmSpecificationV2
+ * @see http://open-services.net/wiki/automation/OSLC-Automation-Specification-Version-2.0/#Resource_AutomationPlan
  */
-public abstract class QmResource
-       extends AbstractResource
+public final class AutomationPlan
+extends AbstractResource
 {
+	private final Set<URI>      contributors                = new TreeSet<URI>();
+    private final Set<URI>      creators                    = new TreeSet<URI>();
     private final Set<URI>      rdfTypes                    = new TreeSet<URI>();
+    private final Set<String>   subjects                    = new TreeSet<String>();
+    private final Set<Property> parameterDefinitions        = new TreeSet<Property>();
 
     private Date     created;
+    private String   description;
     private String   identifier;
     private URI      instanceShape;
     private Date     modified;
     private URI      serviceProvider;
     private String   title;
 
-    public QmResource()
-     {
-         super();
+	public AutomationPlan()
+	{
+		super();
 
-         rdfTypes.add(getRdfType());
-     }
+		rdfTypes.add(URI.create(AutomationConstants.TYPE_AUTOMATION_PLAN));
+	}
 
-     public QmResource(final URI about)
+    public AutomationPlan(final URI about)
      {
          super(about);
 
-         rdfTypes.add(getRdfType());
+		rdfTypes.add(URI.create(AutomationConstants.TYPE_AUTOMATION_PLAN));
      }
 
-    protected abstract URI getRdfType();
+    protected URI getRdfType() {
+    	return URI.create(AutomationConstants.TYPE_AUTOMATION_PLAN);
+    }
+
+    public void addContributor(final URI contributor)
+    {
+        this.contributors.add(contributor);
+    }
+
+    public void addCreator(final URI creator)
+    {
+        this.creators.add(creator);
+    }
 
     public void addRdfType(final URI rdfType)
     {
         this.rdfTypes.add(rdfType);
+    }
+
+    public void addSubject(final String subject)
+    {
+        this.subjects.add(subject);
+    }
+
+    public void addParameterDefinition(final Property parameter)
+    {
+        this.parameterDefinitions.add(parameter);
+    }
+
+    @OslcDescription("The person(s) who are responsible for the work needed to complete the automation plan.")
+    @OslcName("contributor")
+    @OslcPropertyDefinition(OslcConstants.DCTERMS_NAMESPACE + "contributor")
+    @OslcRange(QmConstants.TYPE_PERSON)
+    @OslcTitle("Contributors")
+    public URI[] getContributors()
+    {
+        return contributors.toArray(new URI[contributors.size()]);
     }
 
     @OslcDescription("Timestamp of resource creation.")
@@ -77,6 +119,25 @@ public abstract class QmResource
     public Date getCreated()
     {
         return created;
+    }
+
+    @OslcDescription("Creator or creators of resource.")
+    @OslcName("creator")
+    @OslcPropertyDefinition(OslcConstants.DCTERMS_NAMESPACE + "creator")
+    @OslcRange(QmConstants.TYPE_PERSON)
+    @OslcTitle("Creators")
+    public URI[] getCreators()
+    {
+        return creators.toArray(new URI[creators.size()]);
+    }
+
+    @OslcDescription("Descriptive text (reference: Dublin Core) about resource represented as rich text in XHTML content.")
+    @OslcPropertyDefinition(OslcConstants.DCTERMS_NAMESPACE + "description")
+    @OslcTitle("Description")
+    @OslcValueType(ValueType.XMLLiteral)
+    public String getDescription()
+    {
+        return description;
     }
 
     @OslcDescription("A unique identifier for a resource. Assigned by the service provider when a resource is created. Not intended for end-user display.")
@@ -125,6 +186,16 @@ public abstract class QmResource
         return serviceProvider;
     }
 
+    @OslcDescription("Tag or keyword for a resource. Each occurrence of a dcterms:subject property denotes an additional tag for the resource.")
+    @OslcName("subject")
+    @OslcPropertyDefinition(OslcConstants.DCTERMS_NAMESPACE + "subject")
+    @OslcReadOnly(false)
+    @OslcTitle("Subjects")
+    public String[] getSubjects()
+    {
+        return subjects.toArray(new String[subjects.size()]);
+    }
+
     @OslcDescription("Title (reference: Dublin Core) or often a single line summary of the resource represented as rich text in XHTML content.")
     @OslcOccurs(Occurs.ExactlyOne)
     @OslcPropertyDefinition(OslcConstants.DCTERMS_NAMESPACE + "title")
@@ -135,9 +206,46 @@ public abstract class QmResource
         return title;
     }
 
+    @OslcDescription("The parameter definitions for the automation plan.")
+    @OslcOccurs(Occurs.ZeroOrMany)
+    @OslcName("parameterDefinition")
+    @OslcPropertyDefinition(AutomationConstants.AUTOMATION_NAMESPACE + "parameterDefinition")
+    @OslcValueType(ValueType.LocalResource)
+    @OslcTitle("Parameter Definitions")
+    public Property[] getParameterDefinitions()
+    {
+        return parameterDefinitions.toArray(new Property[parameterDefinitions.size()]);
+    }
+
+
+    public void setContributors(final URI[] contributors)
+    {
+        this.contributors.clear();
+
+        if (contributors != null)
+        {
+            this.contributors.addAll(Arrays.asList(contributors));
+        }
+    }
+
     public void setCreated(final Date created)
     {
         this.created = created;
+    }
+
+    public void setCreators(final URI[] creators)
+    {
+        this.creators.clear();
+
+        if (creators != null)
+        {
+            this.creators.addAll(Arrays.asList(creators));
+        }
+    }
+
+    public void setDescription(final String description)
+    {
+        this.description = description;
     }
 
     public void setIdentifier(final String identifier)
@@ -170,9 +278,29 @@ public abstract class QmResource
         this.serviceProvider = serviceProvider;
     }
 
+    public void setSubjects(final String[] subjects)
+    {
+        this.subjects.clear();
+
+        if (subjects != null)
+        {
+            this.subjects.addAll(Arrays.asList(subjects));
+        }
+    }
+
     public void setTitle(final String title)
     {
         this.title = title;
+    }
+
+    public void setParameterDefinitions(final Property[] parameterDefinitions)
+    {
+        this.parameterDefinitions.clear();
+
+        if (parameterDefinitions != null)
+        {
+            this.parameterDefinitions.addAll(Arrays.asList(parameterDefinitions));
+        }
     }
 
 }
