+++ oslc4j-client/src/main/java/org/eclipse/lyo/client/resources/ChangeRequest.java
@@ -16,7 +16,7 @@
  *     Chris Peters         - initial API and implementation
  *     Gianluca Bernardini  - initial API and implementation
  *******************************************************************************/
-package org.eclipse.lyo.client.oslc.resources;
+package org.eclipse.lyo.client.resources;
 
 import java.net.URI;
 import java.net.URISyntaxException;
