+++ oslc4j-client/src/test/java/org/eclipse/lyo/client/test/OslcQueryResultTest.java
@@ -25,10 +25,11 @@ import java.net.MalformedURLException;
 
 
 import javax.ws.rs.core.Response;
-import org.eclipse.lyo.client.oslc.OslcClient;
-import org.eclipse.lyo.client.oslc.resources.OslcQuery;
-import org.eclipse.lyo.client.oslc.resources.OslcQueryParameters;
-import org.eclipse.lyo.client.oslc.resources.OslcQueryResult;
+
+import org.eclipse.lyo.client.OslcClient;
+import org.eclipse.lyo.client.resources.OslcQuery;
+import org.eclipse.lyo.client.resources.OslcQueryParameters;
+import org.eclipse.lyo.client.resources.OslcQueryResult;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
